refactor: put note-button's components into separate files
